build: add crate "toyc-main" with "main" fn and update "toyc" to just call that
It's not really an issue now, but this should improve compile times.
Rust does not yet support compiling and linking of binary crates as
separate steps, this effectively makes the final binary crate very
cheap to compile so every crate can be compiled as soon as its
dependencies' metadata is ready.